Refactor settings file models to storage module
The Pydantic models for the YAML settings files in a GitHub repo were
part of the domain model for a GitHub Repo checkout but they're more
properly considered part of the storage layer since they're interfaces
to external systems (i.e., we can't change these models without
affecting users).
