+++ src/timessquare/storage/github/__init__.py
@@ -1,0 +1,1 @@
+"""Interfaces for GitHub-based data storage and API operations."""
